requirements-test.txt
-r requirements.txt
pytest
pytest-cov
pytest-django
freezegun
pyquery